pyproject.toml
[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"

[project]
name = "mini-gpt"
version = "0.1.0"
description = "Mini Transformer written from scratch for learning purposes."
authors = [
    {name = "Ki KIm"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "torch (>=2.9.1,<3.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
